Enviando o desafio da aula 03

+++ exercicio08.py
@@ -1,4 +1,4 @@
-numero = float(input("Informe um número: "))
+numero = int(input("Informe um número: "))
 
 resto = numero % 2
